@@ -3,48 +3,48 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.7.1",
+    "@reduxjs/toolkit": "^1.9.5",
     "@stomp/stompjs": "^6.1.2",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "add": "^2.0.6",
-    "axios": "^0.24.0",
+    "axios": "^1.4.0",
     "bootstrap": "^5.1.3",
     "file-saver": "^2.0.5",
-    "firebase": "^9.8.1",
-    "moment": "^2.29.1",
+    "firebase": "^9.23.0",
+    "gradient-string": "^2.0.2",
+    "moment": "^2.29.4",
     "net": "^1.0.2",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-audio-player": "^0.17.0",
-    "react-bootstrap": "^2.2.3",
-    "react-dom": "^17.0.2",
+    "react-bootstrap": "^2.8.0",
+    "react-dom": "^18.2.0",
     "react-draft-wysiwyg": "^1.14.7",
-    "react-ga4": "^1.4.1",
-    "react-icons": "^4.3.1",
-    "react-infinite-scroll-hook": "^4.0.1",
-    "react-moment": "^1.1.1",
-    "react-notifications": "^1.7.3",
-    "react-overlay-component": "^1.0.2",
-    "react-quill": "^1.3.5",
+    "react-ga4": "^2.1.0",
+    "react-icons": "^4.10.1",
+    "react-infinite-scroll-hook": "^4.1.1",
+    "react-moment": "^1.1.3",
+    "react-notifications": "^1.7.4",
+    "react-overlay-component": "^1.0.6",
+    "react-quill": "^2.0.0",
     "react-redux": "^7.2.6",
-    "react-router-bootstrap": "^0.26.1",
-    "react-router-dom": "6",
-    "react-scripts": "5.0.0",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.14.0",
+    "react-scripts": "^5.0.1",
     "react-simple-image-viewer": "^1.1.1",
     "react-slick": "^0.29.0",
+    "react-spinners": "^0.12.0",
     "redux-devtools-extension": "^2.13.9",
     "redux-logger": "^3.0.6",
-    "redux-saga": "^1.1.3",
-    "redux-thunk": "^2.4.1",
-    "rgt": "^1.0.7",
+    "redux-saga": "^1.2.3",
+    "redux-thunk": "^2.4.2",
     "slick-carousel": "^1.8.1",
-    "sockjs-client": "^1.5.2",
+    "sockjs-client": "^1.6.1",
     "stompjs": "^2.3.3",
-    "styled-components": "^5.3.3",
-    "use-long-press": "^1.1.2",
-    "web-vitals": "^2.1.2",
-    "yarn": "^1.22.17"
+    "styled-components": "^5.3.11",
+    "use-long-press": "^3.1.5",
+    "web-vitals": "^3.3.2"
   },
   "scripts": {
     "start": "react-app-rewired start",
