@@ -2,5 +2,6 @@ module.exports = {
   presets:[
     "@babel/preset-env",
     "@babel/preset-react"
-  ]
+  ],
+  plugins: ["@emotion"]
 }
