@@ -1,4 +1,9 @@
 import {Container} from "react-bootstrap";
+import {useEffect, useState} from "react";
+import {
+  fetchLottoNumberCount,
+  fetchUserCount
+} from "../../../../services/firebase.service";
 
 /***
  *
@@ -9,9 +14,35 @@ import {Container} from "react-bootstrap";
  */
 const LottoTotalStatistics = () => {
 
+  const [totalCount, setTotalCount] = useState(0);
+  const [userCount, setUserCount] = useState(0);
+
+  /**
+   *
+   * 번호 생성 카운트
+   *
+   */
+  const fetchNumber = () => {
+    fetchLottoNumberCount().then(count => setTotalCount(count));
+  }
+
+  const fetchUsers = () => {
+    fetchUserCount().then(count => setUserCount(count));
+  }
+
+  useEffect(() => {
+    fetchNumber();
+    fetchUsers();
+  }, [])
+
   return <>
     <Container className="mt-5">
-        데이터 수집 진행중...
+      <h5>지금까지 총 생성된 번호</h5>
+      <div>{totalCount}</div>
+      <hr/>
+      <h5>생성된 유저수</h5>
+      <div>{userCount}</div>
+      <hr/>
     </Container>
 
   </>
